concept-meta-modeling/requirements.txt
pydantic
pyyaml

pytest